refactor(feature-manager): rename config dir to features
