+++ src/app/modules/verified-content/verified-content.service.ts
@@ -1,7 +1,7 @@
 import { err, ok, Result } from 'neverthrow'
 
 import { AuthType } from '../../../types'
-import { webhooksAndVerifiedContentConfig } from '../../config/feature-manager/webhook-verified-content.config'
+import { webhooksAndVerifiedContentConfig } from '../../config/features/webhook-verified-content.config'
 import formsgSdk from '../../config/formsg-sdk'
 import { createLoggerWithLabel } from '../../config/logger'
