+++ src/app/services/captcha/captcha.service.ts
@@ -1,7 +1,7 @@
 import axios from 'axios'
 import { errAsync, okAsync, ResultAsync } from 'neverthrow'
 
-import { captchaConfig } from '../../config/feature-manager/captcha.config'
+import { captchaConfig } from '../../config/features/captcha.config'
 import { createLoggerWithLabel } from '../../config/logger'
 
 import { GOOGLE_RECAPTCHA_URL } from './captcha.constants'
